+++ yandex/resource_yandex_compute_instance.go
@@ -512,9 +512,8 @@ func resourceYandexComputeInstance() *schema.Resource {
 			},
 
 			"local_disk": {
-				Type:         schema.TypeList,
-				Optional:     true,
-				RequiredWith: []string{"placement_policy"},
+				Type:     schema.TypeList,
+				Optional: true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"size_bytes": {