+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 BUG FIXES:
 * mysql: fix bug leading to cluster recreation when changing folder_id
 * clickhouse: fix bug leading to failed updating cluster when set cloud_storage(enabled: false)
+* compute: fix validation of local_disk in yandex_compute_instance resource
 
 ## 0.83.0 (December 1, 2022)
 BUG FIXES: